#78 Error Protocols and Refactoring More Settings
