+++ TrailblazerIOS/Views/MoreSettingsViews/MoreSettingsDocumentPickerDelegate.swift
@@ -1,0 +1,30 @@
+//
+//  MoreSettingsDocumentPickerDelegate.swift
+//  Trailblazer
+//
+//  Created by Peyton McKee on 3/14/23.
+//
+
+import Foundation
+import UIKit
+
+extension MoreSettingViewController: UIDocumentPickerDelegate {
+    public func documentPicker(_ controller: UIDocumentPickerViewController, didPickDocumentsAt urls: [URL]) {
+        guard let myURL = urls.first else {
+            return
+        }
+        print("import result : \(myURL)")
+        do {
+            let data = try String(contentsOf: myURL)
+            self.mapFile = data
+        }
+        catch{
+            print("Could not get contents of file")
+        }
+    }
+    
+    func documentPickerWasCancelled(_ controller: UIDocumentPickerViewController) {
+        print("view was cancelled")
+        self.dismiss(animated: true, completion: nil)
+    }
+}
